+++ packages/nextjs/src/components/Link.tsx
@@ -9,6 +9,19 @@ import {
 } from '@sitecore-content-sdk/react';
 
 /**
+ * The list of NextLink props to be supported by the Link component.
+ */
+const supportedNextLinkProps = [
+  'as',
+  'onNavigate',
+  'passHref',
+  'prefetch',
+  'replace',
+  'scroll',
+  'shallow',
+] as const;
+
+/**
  * The interface for the Link component props.
  * @public
  */
@@ -18,11 +31,7 @@ export type LinkProps = ReactLinkProps & {
    * @default /^\//g
    */
   internalLinkMatcher?: RegExp;
-  /**
-   * Next.js Link prefetch.
-   */
-  prefetch?: NextLinkProps['prefetch'];
-};
+} & Pick<NextLinkProps, (typeof supportedNextLinkProps)[number]>;
 
 /**
  * Matches relative URLs that end with a file extension.
@@ -41,7 +50,7 @@ export const Link = forwardRef<HTMLAnchorElement, LinkProps>(
       children,
       internalLinkMatcher = /^\//g,
       showLinkTextWithChildrenPresent,
-      ...htmlLinkProps
+      ...rest
     } = props;
 
     if (!field || (!field.value && !(field as LinkFieldValue).href && !field.metadata)) {
@@ -68,12 +77,11 @@ export const Link = forwardRef<HTMLAnchorElement, LinkProps>(
           <NextLink
             href={{ pathname: href, query: querystring, hash: anchor }}
             key="link"
-            locale={false}
             title={value.title}
             target={value.target}
             className={value.class}
-            prefetch={props.prefetch}
-            {...htmlLinkProps}
+            {...rest}
+            locale={false}
             ref={ref}
             {...(process.env.TEST
               ? {
@@ -89,10 +97,7 @@ export const Link = forwardRef<HTMLAnchorElement, LinkProps>(
       }
     }
 
-    // prevent passing internalLinkMatcher or prefetch as it is an invalid DOM element prop
-    const reactLinkProps = { ...props };
-    delete reactLinkProps.internalLinkMatcher;
-    delete reactLinkProps.prefetch;
+    const reactLinkProps = sanitizeLinkProps(props);
 
     return (
       <ReactLink
@@ -105,3 +110,29 @@ export const Link = forwardRef<HTMLAnchorElement, LinkProps>(
 );
 
 Link.displayName = 'NextLink';
+
+/**
+ * Sanitize props for ReactLink by removing Next.js and internal props to prevent invalid DOM attributes.
+ * @param {LinkProps} props - The props the Link component received.
+ * @returns sanitized props for ReactLink.
+ * @internal
+ */
+function sanitizeLinkProps(props: LinkProps) {
+  const nextLinkProps: (keyof NextLinkProps)[] = [
+    'as',
+    'onNavigate',
+    'passHref',
+    'prefetch',
+    'replace',
+    'scroll',
+    'shallow',
+  ];
+  const internalProps: (keyof LinkProps)[] = ['internalLinkMatcher'];
+
+  const sanitizedProps: LinkProps = { ...props };
+  for (const prop of [...nextLinkProps, ...internalProps]) {
+    delete sanitizedProps[prop as keyof LinkProps];
+  }
+
+  return sanitizedProps;
+}